refactor: reorganize bank composables and update imports to align with new structure
