+++ app/src/composables/bank/writes.ts
@@ -1,199 +1,171 @@
 import { computed, unref, type MaybeRef } from 'vue'
+import type { Address } from 'viem'
 import { useQueryClient } from '@tanstack/vue-query'
-import { useChainId } from '@wagmi/vue'
-import { useTeamStore, useUserDataStore } from '@/stores'
-
-import { BANK_FUNCTION_NAMES, type BankFunctionName, isValidBankFunction } from './types'
-
-import {
-  useContractWrites,
-  type ContractWriteConfig,
-  type ContractWriteOptions
-} from '../contracts/useContractWrites'
 import { BANK_ABI } from '@/artifacts/abi/bank'
+import { useContractWrites } from '@/composables/contracts/useContractWritesV2'
+import { useTeamStore } from '@/stores/teamStore'
+import type { ExtractAbiFunctionNames } from 'abitype'
 
-/**
- * Bank contract specific write operations
- * This is a wrapper around the generic useContractWrites composable
- */
-export function useBankWrites() {
+type BankFunctionNames = ExtractAbiFunctionNames<typeof BANK_ABI>
+
+// Helper function to wrap useContractWrites for Bank contract
+export function useBankContractWrite(options: {
+  functionName: BankFunctionNames
+  args?: MaybeRef<readonly unknown[]>
+  value?: MaybeRef<bigint>
+}) {
   const teamStore = useTeamStore()
-  const queryClient = useQueryClient()
-  const chainId = useChainId()
   const bankAddress = computed(() => teamStore.getContractAddressByType('Bank'))
-  const userDataStore = useUserDataStore()
-  const userAddr = userDataStore.address as `0x${string}` | undefined
-  // Use the generic contract writes composable
-  const baseWrites = useContractWrites({
-    contractAddress: bankAddress.value!,
+  return useContractWrites({
+    contractAddress: bankAddress,
     abi: BANK_ABI,
-    chainId: chainId.value
-  } as ContractWriteConfig)
+    functionName: options.functionName,
+    args: options.args ?? [],
+    ...(options.value !== undefined ? { value: options.value } : {})
+  })
+}
 
-  /**
-   * Bank-specific query invalidation based on function name
-   */
-  const invalidateBankQueries = async (
-    functionName: BankFunctionName,
-    args?: readonly unknown[]
-  ) => {
-    if (!bankAddress.value) return
+export function useDepositToken(token: MaybeRef<Address>, amount: MaybeRef<bigint>) {
+  const args = computed(() => [unref(token), unref(amount)] as readonly unknown[])
 
-    const bankQueryKey = {
-      address: bankAddress.value,
-      chainId: chainId.value
-    }
+  return useBankContractWrite({
+    functionName: 'depositToken',
+    args
+  })
+}
 
-    // Invalidate queries based on the function that was called
-    switch (functionName) {
-      case BANK_FUNCTION_NAMES.PAUSE:
-      case BANK_FUNCTION_NAMES.UNPAUSE:
-        // Invalidate paused status query
-        const key = {
-          queryKey: [
-            'readContract',
-            {
-              ...bankQueryKey,
-              functionName: BANK_FUNCTION_NAMES.PAUSED
-            }
-          ]
-        }
-        if (process.env.NODE_ENV === 'development') {
-          console.log('try to invalidate this key', key)
-        }
-        await queryClient.invalidateQueries(key)
-        break
+export function useAddTokenSupport(tokenAddress: MaybeRef<Address>) {
+  return useBankContractWrite({
+    functionName: 'addTokenSupport',
+    args: [tokenAddress]
+  })
+}
 
-      case BANK_FUNCTION_NAMES.TRANSFER_OWNERSHIP:
-      case BANK_FUNCTION_NAMES.RENOUNCE_OWNERSHIP:
-        // Invalidate owner query
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'readContract',
-            {
-              ...bankQueryKey,
-              functionName: BANK_FUNCTION_NAMES.OWNER
-            }
-          ]
-        })
-        break
+export function useRemoveTokenSupport(tokenAddress: MaybeRef<Address>) {
+  return useBankContractWrite({
+    functionName: 'removeTokenSupport',
+    args: [tokenAddress]
+  })
+}
 
-      case BANK_FUNCTION_NAMES.DEPOSIT_DIVIDENDS:
-        if (userAddr) {
-          // precise: only the current wallet’s dividend balance
-          await queryClient.invalidateQueries({
-            queryKey: [
-              'readContract',
-              {
-                ...bankQueryKey, // { address: bankAddress.value, chainId: chainId.value }
-                functionName: 'dividendBalances',
-                args: [userAddr] as const // mapping key
-              }
-            ]
-          })
-        }
-        break
+export function useClaimDividend() {
+  const queryClient = useQueryClient()
+  const teamStore = useTeamStore()
+  const bankAddress = computed(() => teamStore.getContractAddressByType('Bank'))
+  
+  const writeResult = useBankContractWrite({
+    functionName: 'claimDividend',
+    args: []
+  })
 
-      case BANK_FUNCTION_NAMES.DEPOSIT_TOKEN_DIVIDENDS:
-        if (userAddr && args?.[0]) {
-          // precise: only the current wallet’s dividend balance
-          await queryClient.invalidateQueries({
-            queryKey: [
-              'readContract',
-              {
-                ...bankQueryKey, // { address: bankAddress.value, chainId: chainId.value }
-                functionName: 'tokenDividendBalances',
-                args: [args[0], userAddr] as const // [tokenAddress, userAddress]
-              }
-            ]
-          })
-        }
-        break
-      case BANK_FUNCTION_NAMES.TRANSFER:
-      case BANK_FUNCTION_NAMES.TRANSFER_TOKEN:
-      case BANK_FUNCTION_NAMES.DEPOSIT_TOKEN:
-        // Invalidate balance queries
-        await queryClient.invalidateQueries({
-          queryKey: ['balance', bankQueryKey]
-        })
-        // Also invalidate all readContract queries for this bank address
-        await queryClient.invalidateQueries({
-          queryKey: ['readContract', { ...bankQueryKey, address: args ? args[0] : undefined }]
-        })
-        break
-      default:
-        // For any other function, invalidate all bank-related queries
-        await queryClient.invalidateQueries({
-          queryKey: ['readContract', bankQueryKey]
-        })
-        break
-    }
+  // Override invalidateQueries to invalidate dividend balances
+  const originalInvalidateQueries = writeResult.invalidateQueries
+  writeResult.invalidateQueries = async () => {
+    await originalInvalidateQueries()
+    // Invalidate all dividend balances queries for this bank
+    await queryClient.invalidateQueries({
+      queryKey: ['bank', 'dividendBalances', bankAddress.value]
+    })
   }
 
-  /**
-   * Execute a Bank contract write operation with validation
-   */
-  const executeWrite = async (
-    functionName: BankFunctionName,
-    args: readonly unknown[] = [],
-    value?: MaybeRef<bigint>,
-    options?: ContractWriteOptions
-  ) => {
-    if (!isValidBankFunction(functionName)) {
-      throw new Error(`Invalid bank function: ${functionName}`)
-    }
+  return writeResult
+}
 
-    const result = await baseWrites.executeWrite(
-      functionName,
-      args,
-      value ? unref(value) : undefined,
-      options
-    )
-    await invalidateBankQueries(functionName, args)
-    return result
+export function useClaimTokenDividend(token: MaybeRef<Address>) {
+  const queryClient = useQueryClient()
+  const teamStore = useTeamStore()
+  const bankAddress = computed(() => teamStore.getContractAddressByType('Bank'))
+
+  const args = computed(() => [unref(token)] as readonly unknown[])
+  
+  const writeResult = useBankContractWrite({
+    functionName: 'claimTokenDividend',
+    args
+  })
+
+  // Override invalidateQueries to invalidate dividend balances
+  const originalInvalidateQueries = writeResult.invalidateQueries
+  writeResult.invalidateQueries = async () => {
+    await originalInvalidateQueries()
+    // Invalidate all dividend balances queries for this bank
+    await queryClient.invalidateQueries({
+      queryKey: ['bank', 'dividendBalances', bankAddress.value]
+    })
   }
 
-  /**
-   * Estimate gas for a Bank function call
-   */
-  const estimateGas = async (
-    functionName: BankFunctionName,
-    args: readonly unknown[] = [],
-    value?: MaybeRef<bigint>
-  ) => {
-    if (!isValidBankFunction(functionName)) {
-      throw new Error(`Invalid bank function: ${functionName}`)
-    }
+  return writeResult
+}
 
-    return baseWrites.estimateGas(functionName, args, value ? unref(value) : undefined)
-  }
+export function useDepositDividends(amount: MaybeRef<bigint>, investorAddress: MaybeRef<Address>) {
+  const args = computed(() => [unref(amount), unref(investorAddress)] as readonly unknown[])
+  return useBankContractWrite({
+    functionName: 'depositDividends',
+    args,
+    value: amount // This is a payable function
+  })
+}
 
-  /**
-   * Check if a Bank transaction will likely succeed
-   */
-  const canExecuteTransaction = async (
-    functionName: BankFunctionName,
-    args: readonly unknown[] = [],
-    value?: MaybeRef<bigint>
-  ): Promise<boolean> => {
-    if (!isValidBankFunction(functionName)) {
-      return false
-    }
+export function useDepositTokenDividends(
+  token: MaybeRef<Address>,
+  amount: MaybeRef<bigint>,
+  investorAddress: MaybeRef<Address>
+) {
+  const args = computed(() => [unref(token), unref(amount), unref(investorAddress)] as readonly unknown[])
+  return useBankContractWrite({
+    functionName: 'depositTokenDividends',
+    args
+  })
+}
 
-    return baseWrites.canExecuteTransaction(functionName, args, value ? unref(value) : undefined)
-  }
+export function useSetInvestorAddress(investorAddress: MaybeRef<Address>) {
+  return useBankContractWrite({
+    functionName: 'setInvestorAddress',
+    args: [investorAddress]
+  })
+}
 
-  return {
-    // Re-export all base functionality
-    ...baseWrites,
+export function useTransfer(to: MaybeRef<Address>, amount: MaybeRef<bigint>) {
+  return useBankContractWrite({
+    functionName: 'transfer',
+    args: [to, amount]
+  })
+}
 
-    // Override with Bank-specific implementations
-    executeWrite,
-    estimateGas,
-    canExecuteTransaction,
-    invalidateBankQueries,
+export function useTransferToken(
+  token: MaybeRef<Address>,
+  to: MaybeRef<Address>,
+  amount: MaybeRef<bigint>
+) {
+  return useBankContractWrite({
+    functionName: 'transferToken',
+    args: [token, to, amount]
+  })
+}
 
-    // Keep the generic invalidateQueries as well for flexibility
-    invalidateQueries: baseWrites.invalidateQueries
-  }
+export function useTransferOwnership(newOwner: MaybeRef<Address>) {
+  return useBankContractWrite({
+    functionName: 'transferOwnership',
+    args: [newOwner]
+  })
+}
+
+export function useRenounceOwnership() {
+  return useBankContractWrite({
+    functionName: 'renounceOwnership',
+    args: []
+  })
+}
+
+export function usePause() {
+  return useBankContractWrite({
+    functionName: 'pause',
+    args: []
+  })
+}
+
+export function useUnpause() {
+  return useBankContractWrite({
+    functionName: 'unpause',
+    args: []
+  })
 }
